改用浮点纹理

+++ ClosestGLTests/SurfaceViewer.h
@@ -3,7 +3,7 @@
 #include <Vector4.h>
 namespace ClosestGLTests::Tools
 {
-	using TestCol = ClosestGL::Math::Vector4<uint8_t>;
+	using TestCol = ClosestGL::Math::Vector4<float>;
 	using TestTex = ClosestGL::Texture::Texture2D<TestCol>;
 	void ViewSurface(TestTex& sur,uint64_t time = 5000);
 }
